Syntax fix after merge

+++ common/src/main/kotlin/it/polito/master/ap/group6/ecommerce/common/dtos/ProductDTO.kt
@@ -6,5 +6,5 @@ data class ProductDTO(
     val description: String? = null,
     val category: String? = null,
     val picture: String? = null,
-    val currentPrice: Double? = null
+    val currentPrice: Float? = null
 )
